Process review comments on 260
