--- verifier/src/test/java/com/aws/jverify/verifier/tests/Blep.java
@@ -1,4 +1,0 @@
-package com.aws.jverify.verifier.tests;
-
-public class Blep {
-}
